Avoid `invalidConfigPath` for function values
Pass `validateConfigPath` if a type of value equals to `function`. The purpose of this commit is to avoid unwanted warnings when we define theme values as functions in tailwind.config.js, which, for example, are required to define theme colors with opacity enabled. https://github.com/tailwindlabs/tailwindcss/discussions/2206#discussioncomment-60081