Collect `Scope` metadata about document content
Here we introduce the concept of a `Scope` which is an abstract marker for a region of content in a document. A `Scope` can be used to mark a region of text with specific metadata that can be used to drive various features in the editor. For example, a `Scope` might be used to represent things like whether a region of text is HTML or CSS; whether or not we're inside a comment, string literal, or other language-specific construct; or whether or not we're looking at a class list (regardless of the language), etc…