Fix suggestion of utilities with slashes in them in v4 (#1182)
All utilities implicitly have at least `modifiers: []` in the class list
when no modifiers are present. We should only change the completion list
when there are actually modifiers to show in case an existing utility
with a slash in it (e.g. `w-1/2`) is “close” to what the user has typed.

This is the other half to fix #1178 b/c while the completions show up
but as soon as you type `/` they disappear which is an awful experience.