Always use `1rem = 16px` for pixel equivalents in media queries (#1190)
Fixes #1189

From [the CSS spec](https://drafts.csswg.org/mediaqueries/#units):

> Relative length units in media queries are based on the initial value,
which means that units are never based on results of declarations. For
example, in HTML, the em unit is relative to the initial value of
font-size, defined by the user agent or the user’s preferences, not any
styling on the page.

We should hide these because `1rem` everywhere else will be one value
but `1rem` in a media query might still be 16px (it's not changeable by
websites — only browsers / users)