Fix detection when project contains stylesheets that import the "main" stylesheet (#1218)
Fixes #1217

This issue kinda reveals that the entire detection around "project
roots" or "root stylesheets" needs to be rethought because files at the
top of a CSS graph may not actually be the intended stylesheet —
especially when `@reference` is used. But… like anything… they also
_might be_.

In the meantime we'll "prefer" files we can detect as definite root
stylesheets before trying other files.