v4: Support loading bundled versions of some first-party plugins (#1240)
see
https://github.com/tailwindlabs/tailwindcss-intellisense/issues/1224#issuecomment-2687637180

This adds support for loading these three plugins in v4 **only when
using `@plugin`** if we can't find them:
- `@tailwindcss/typography`
- `@tailwindcss/forms`
- `@tailwindcss/aspect-ratio`

This coincides with behavior of the Standalone CLI where these are
bundled instead of available as an NPM package.

I've additionally logs when trying to import one of these plugins inside
a JS file. This does not work right now and may take some effort to
support because it will at least require the use of Node's experimental
loader's API and potentially additional work as well.