Cancel initial file search if it takes too long (#1242)
Fixes #986 (hopefully)

There's a lot of detail about the problem in the linked issue and my
response at the bottom. But briefly:
- Searches can take way too long; even spinning "forever" for some
people
- This occurs more often if following symlinks is enabled in VSCode
- We can cancel in-flight searches

This PR starts the searches and cancels them after ~15s. If the
extension hasn't found anything within 15s then it's likely it'll take
~15s any time you open that project and that would've been an absolutely
horrible dev experience that we probably would've been notified about.

It seems reasonable to set a time limit on the search do determine if we
need to start a language server