LSP: Refresh internal caches when settings are updated (#1273)
We use the pull model (`workspace/configuration`) to get settings for a
document and we cache these internally so we don't repeat these calls
multiple times for a given request. We're set up to listen for
configuration refresh notifications and update the project settings when
we get them.

Unfortunately, we didn't actually _register_ for these notifications, so
we never got them. This meant that if you changed the settings for an
already opened file or workspace folder, the language server would not
react to these changes. This PR fixes this by registering for
configuration change notifications and now open files with color
decorators, completions, etc… should react to changes in the settings as
needed.

If settings are updated and our langauge server doesn't react to or
handle these changes, it is definitely a bug. Hopefully this will squash
all of those particular ones but… we'll see. 😅