Improve error message when a workspace folder is inaccessible (#1276)
On Linux when using a LSP client that does not support watching files on
behalf of the server, we'll use Parcel Watcher (if possible). If we
start the watcher with a non-existent or inaccessible directory, it will
throw an error with a very unhelpful message: "Bad file descriptor".

The best thing we can do is an initial check for access to the directory
and log a more helpful error message if it fails:

<img width="1013" alt="Screenshot 2025-03-20 at 12 33 29"
src="https://github.com/user-attachments/assets/425c1fac-cd5e-492f-89bf-76d3a733670d"
/>

See
https://github.com/tailwindlabs/tailwindcss-intellisense/issues/884#issuecomment-1927660975

---------

Co-authored-by: Robin Malfait <malfait.robin@gmail.com>