Fix problem with too many ripgrep processes being spawned by VSCode (#1287)
We run a file search on a per-workspace basis but we wait until a file
has been opened. However, there are two big problems here:

- A file being "opened" does **not** mean it is visible. Just that an
extension has, effectively, taken an interest in it, can read its
contents, etc. This happens for things like tsconfig files, some files
inside a `.git` folder, etc…

- We're running the search any time we see an opened document. What
should happen is that we run the search when a document is opened _and
visible_, the language server has not started, and we need to checking a
workspace folder that has not been searched yet.

This code here needs to be restructured to ensure that these searches
only run when they are needed. If the searches don't return anything or
time out then they should not be run again. Notifications from file
watching should take care of the rest in case the initial search turned
up nothing and the user adds a file that should cause the server to
start.

Fixes #986 (for real maybe this time??)