Don't follow recursive symlinks during project discovery (#1270)
Project discovery is done by crawling the workspace with `fast-glob`
looking for CSS files, JS/TS config files, Sass/Less files, etc…

A major problem with this approach though is that `fast-glob` crawls the
filesystem and traverses into symlinked directories. If the symlink
points back to a parent folder the crawler may cause a handful of
issues:
- An infinite traversal loop
- Out of memory crashes (causing the server to restart repeatedly)
- A rather long discovery time (10s+)
- Errors because of file path length limits (esp on Windows)

I've rewritten the traversal to use `tinyglobby` which is a
`fast-glob`-compatible replacement. It avoids all of these issues due to
its use of `fdir` which explicitly detects and avoids traversal of
recursive symlinks.

All existing tests pass as well as an additional test that previously
timed out due to the use of recursive symlinks. This PR is going to
require extensive testing on Windows with network shares, mapped drives,
etc…

Fixes #1056
Fixes #1185