Add test for `classAttributes` matching variable names in JS(X)/TS(X) (#1315)
Welp… this was 100% a bug that this behavior got introduced but at this
point it’s shipped and people might rely on it or otherwise want this
behavior. As such I’m explicitly encoding this in a test so it doesn’t
break later when the big parsing refactor lands.