Support class function hovers in Svelte and HTML `<script>` blocks (#1311)
When using the `tailwindCSS.classFunctions` setting we were handling
completions correctly inside `<script>` tags but not class hovers. This
is b/c the code path for the two is just… different.

It'll get unified eventually (soon hopefully) but until then this is the
bug fix to ensure that hovering inside something like `<script>let
classes = clsx("flex relative")</script>` works