Fix matching files when config is not in the workspace root (#1412)
In IntelliSense we have this special behavior where if we only detect a
single config in a workspace then that config will cover files opened in
the entire workspace. We broke this accidentally in v0.14.16 because we
made sure document selectors were re-calculated correctly for v4
projects files. With how things are structured/shared this setup is the
same for v3 and v4 and we accidentally wiped out the "root" selector
we'd add to the project if it was the only one in a given workspace
folder.

So this isn't *usually* a problem but definitely can be if:
- You have a v3 project where the config file isn't in the root
- You have a v4 project where automatic source detection was disabled or
failed to run

This should hopefully fix any project matching issues people have had
that broke from v0.14.17 and later.

Fixes #1404 (and maybe #1377)

- [x] tests