Suggest default variant values when they also support arbitrary values (#1439)
Given the following plugin:
```js
matchVariant('foo', (value) => ':is(' + value + ')', {
  values: {
    DEFAULT: 'default',
    a: 'a',
  }
})
```

We wouldn't suggest `foo:` as a variant even though it works. This is
because we were suggesting the arbitrary `foo-[]:` *or* the default
variant instead of both if a variant supported arbitrary values in v4.

This PR fixes the above and now both will be suggested.