Fix edit range being incorrect when utilities contain variants (#1480)
Fixes #1479

This is the result of a parsing issue inside core which causes us to
make incorrect assumptions. See
https://github.com/tailwindlabs/tailwindcss/pull/19172

The tl;dr of how this bug happened:
1. We see `hover:bg-[` and each "part" gets tested
2. To test each part we compile `{part}:[color:red]`
3. We test `hover` and it passes — awesome ✅ 
4. We test `bg-[` and it passes — bad ❌
5. This is because it got turned into `bg-[:[color:red]` and was
incorrectly parsed as a valid utility. This is fixed in Tailwind CSS
v4.1.16

This meant that the start position was *after* the end position for a
completion's edit range because it assumed there was a separator (`:`)
after the `bg-[` part when there is not.

We'll work around this for old versions. As mentioned new Tailwind CSS
versions will handle this appropriately.