Tweak internal CSS AST implementation (#1512)
Right now our CSS AST doesn't differentiate at-rules with an empty body
and at-rules that have no body at all.

This is normally not a problem _however_ for CSS mixins `@apply --foo;`
and `@apply --foo {};` mean different things — this requires us to be
able to differentiate them. We'll be changing this in Tailwind CSS
itself at some point so I'm changing this here to be forward compatible
with those changes.